Refactor db utils and add ingestion handler test
